refractored the predicates choosen into two parts. - a seperate method for the menu - added a PredicateEvenrtArgs who hold the choice - fired the onPredicteChosen in the menu method after a person made a choice
